contract: rename into operation module
