+++ src/operation/fungible.rs
