+++ src/operation/global.rs
